arm: dts: suniv: Add empty mmc2 node to make generic sunxi happy
Most (if not all?) Allwinner SoCs except for the F1C100s family have at
least three MMC controllers: mmc0, 1 and 2; where the third one is the
eMMC controller, which the F1C100s lacks.

For fastboot reasons, the mmc2 controller is unconditionnally aliased
to mmc1 in sunxi-u-boot.dtsi, which fails the compilation of suniv-based
socs, as they have no mmc2.

This patch adds an empty mmc2 node to the device tree to make the
compiler happy.

Fixes: (708b5da38) arm: dts: sunxi: Alter mmc2 auto-numbering to mmc1